requirements.txt
langchain
transformers
torch
faiss-cpu
tiktoken
PyPDF2
llama-index
rwkv
numpy
rwkv
